refactor: Restructure graph.rs
